Movement and sample input components

+++ src/PhysicsComponent.h
@@ -8,7 +8,7 @@
 
 class PhysicsComponent : public Component{
 public:
-	explicit PhysicsComponent(int speed);
+	explicit PhysicsComponent(float speed);
 
 	void update(int frameTime);
 
@@ -16,9 +16,9 @@ public:
 	sfld::Vector2f getDirection() const;
 
 	void setDirection(const sfld::Vector2f& unnormalisedVector);
-	void setSpeed(int speed);
+	void setSpeed(float speed);
 private:
 	sfld::Vector2f direction_;
 	sfld::Vector2f velocity_;
-	int speed_;
+	float speed_;
 };
